requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
boto3==1.17.45
botocore==1.20.45
certifi==2020.6.20
cffi==1.14.3
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.14
django-autoslug==1.9.8
django-cors-headers==3.11.0
django-crispy-forms==1.11.2
django-file-validator==0.0.4
django-generic-follow==0.4.2
django-gravatar2==1.4.4
django-haystack==3.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-restframework==0.0.1
django-storages==1.11.1
django-user-visit==0.4
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
dnspython==2.1.0
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jmespath==0.10.0
msgpack==1.0.2
oauthlib==3.1.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.6
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
Twisted==21.2.0
twisted-iocpsupport==1.0.4
txaio==20.12.1
ua-parser==0.10.0
urllib3==1.25.10
user-agents==2.2.0
whitenoise==5.2.0
zope.interface==5.2.0
